Release 1.0.0 - Calculating altitude and converting coordinates

+++ main.js
@@ -1,15 +1,47 @@
-const { app, BrowserWindow, ipcMain } = require('electron')
+const { app, BrowserWindow, BaseWindow, dialog, ipcMain } = require('electron')
+if(require('electron-squirrel-startup')) app.quit();
 const path = require('node:path')
 const { spawn } = require('child_process')
+const fs = require('fs')
+const admZip = require("adm-zip");
 
 let elevatorServer = null;
-let height = null;
+const SERVER_PATH = './resources/elevation_server.exe'
+const DEM_TILES_PATH = './resources/dem_tiles'
+
+const createDemTiles = () => {
+    if (fs.existsSync(DEM_TILES_PATH) && fs.existsSync(DEM_TILES_PATH + '.idx')) return true
+
+    dialog.showMessageBoxSync({
+        title: 'Необходима база данных высот над уровнем моря',
+        message: "Укажите архив 'demTiles.zip'"
+    })
+
+    const arrayPath = dialog.showOpenDialogSync({
+        filters: [
+            { name: 'demTiles', extensions: ['zip'] }
+        ],
+        properties: ['openFile'],
+    })
+    if (!arrayPath || arrayPath[0].includes('demTiles.zip') === false) return false
+
+    const zip = new admZip(arrayPath[0])
+    zip.extractAllTo("./resources", true)
+
+    return fs.existsSync(DEM_TILES_PATH) && fs.existsSync(DEM_TILES_PATH + '.idx');
+}
 
 const createServer = () => {
-    elevatorServer = spawn('server/elevation_server.exe', [
-        '-dem',
-        path.join(__dirname, 'server/dem_tiles')
-    ])
+    elevatorServer = spawn(
+        SERVER_PATH,
+        [
+            '-dem',
+            DEM_TILES_PATH
+        ],
+        {
+            env: {"ELECTRON_RUN_AS_NODE": "1"}
+        }
+    )
 
     elevatorServer.stdout.on("data",function(data){
         console.log("ElevatorServer Data: " + data)
@@ -27,8 +59,6 @@ const createServer = () => {
 }
 
 const createWindow = () => {
-    createServer();
-
     const mainWindow = new BrowserWindow({
         width: 800,
         height: 600,
@@ -37,46 +67,37 @@ const createWindow = () => {
         }
     })
 
-    ipcMain.on('set-request', (event, coordinates) => {
-        const request = spawn('powershell.exe', [
-            path.join(__dirname, 'server/request.ps1'),
-            `${coordinates.latitude}`,
-            `${coordinates.longitude}`
-        ])
-
-        request.stdout.on("data",function(data){
-            console.log("Request Script Data: " + data)
-            height = Number(data)
-        })
-
-        request.stderr.on("data",function(data){
-            console.log("Request Script Errors: " + data)
-        })
-
-        request.on("exit",function(){
-            console.log("Request Script finished")
-        })
-
-        request.stdin.end()
+    ipcMain.on('load-db', () => {
+        if (!createDemTiles()) {
+            dialog.showErrorBox('Нет базы данных высот', 'Приложение будет закрыто!')
+            app.quit()
+            return
+        }
+        if (!elevatorServer) createServer()
     })
 
-    ipcMain.handle('get-response', async () => height)
+    ipcMain.handle('wait-db-loaded', async () => {
+        const checkLoaded = () => {
+            if (elevatorServer === null) {
+                setTimeout(checkLoaded, 1000)
+            }
+        }
+        checkLoaded()
+        return true
+    })
 
     mainWindow.loadFile('./resources/index.html').then()
 }
 
 app.on('ready', () => {
     createWindow()
-
     app.on('activate', () => {
         if (BrowserWindow.getAllWindows().length === 0) createWindow()
     })
 })
 
 app.on('window-all-closed', () => {
-
-    elevatorServer.kill('SIGINT')
-
+    if (elevatorServer) elevatorServer.kill('SIGINT')
     if (process.platform !== 'darwin') {
         app.quit()
     }
